refactor container and day js init logic
