+++ src/lib/dayjsInstance.ts
@@ -1,0 +1,9 @@
+import dayjs from "dayjs"
+
+import timezonePlugin from "dayjs/plugin/timezone"
+import utcPlugin from "dayjs/plugin/utc"
+
+dayjs.extend(utcPlugin)
+dayjs.extend(timezonePlugin)
+
+export const dayjsInstance = dayjs
