+++ src/lib/setup.ts
@@ -3,6 +3,7 @@ process.env.NODE_ENV ??= "development";
 
 import { ApplicationCommandRegistries, RegisterBehavior } from "@sapphire/framework";
 import "@sapphire/plugin-api/register";
+
 import "@sapphire/plugin-editable-commands/register";
 import "@sapphire/plugin-logger/register";
 import "@sapphire/plugin-subcommands/register";
@@ -11,6 +12,7 @@ import * as colorette from "colorette";
 import { join } from "path";
 import { inspect } from "util";
 import { srcDir } from "./constants";
+import "./containerInit"
 
 // Set default behavior to bulk overwrite
 ApplicationCommandRegistries.setDefaultBehaviorWhenNotIdentical(RegisterBehavior.BulkOverwrite);
